requirements.txt
iso8601
kafka-python>=0.9.1,<0.9.3
monasca-statsd
python-statsd>=1.6.3
PyYAML
MySQL-python>=1.2.3
oslo.config>=1.2.1
oslo.utils
pbr>=0.6,!=0.7,<1.0
